New hsl math to fix (big) calc errors. Needs add'l testing.

+++ hsl.h
@@ -9,7 +9,7 @@ struct Rgb {
 
 struct Hsl {
   int hue;
-  byte saturation;
+  float saturation;
   float lightness;
 };
